tools/requirements.txt
# Create a Python virtual environment and install the dependencies listed below
# by running the following commands:
#
# - Windows (using PowerShell):
#     py -m venv env
#     env\Scripts\Activate.ps1
#     py -m pip install -r tools\requirements.txt
#
# - Windows (using Cygwin/MSys2), Linux or macOS:
#     python3 -m venv env
#     source env/bin/activate
#     pip3 install -r tools/requirements.txt

av     >= 15.0.0
Cython >= 3.1.3
numpy  >= 2.3.2
Pillow >= 11.3.0
